+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/AdjacentPluginCompatTest.kt
@@ -70,7 +70,7 @@ class AdjacentPluginCompatTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("com.vanniktech.maven.publish") version "0.30.0"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }