+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/BuildTimeBenchmarkTest.kt
@@ -59,7 +59,7 @@ class BuildTimeBenchmarkTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {