+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/EdgeCaseMatrixTest.kt
@@ -76,7 +76,7 @@ class EdgeCaseMatrixTest {
         buildFile.writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin { jvm("desktop") }
@@ -108,7 +108,7 @@ class EdgeCaseMatrixTest {
         buildFile.writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin { jvm("desktop") }
@@ -141,7 +141,7 @@ class EdgeCaseMatrixTest {
         buildFile.writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin { jvm("desktop") }