+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/FlavorMainSourceSetLivenessTest.kt
@@ -47,7 +47,7 @@ class FlavorMainSourceSetLivenessTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin { jvm("desktop") }