+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/GaReadinessIntegrationTest.kt
@@ -83,7 +83,7 @@ class GaReadinessIntegrationTest {
         File(projectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {
@@ -125,7 +125,7 @@ class GaReadinessIntegrationTest {
         File(projectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {
@@ -157,7 +157,7 @@ class GaReadinessIntegrationTest {
         File(projectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {
@@ -206,7 +206,7 @@ class GaReadinessIntegrationTest {
         File(projectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {
@@ -252,7 +252,7 @@ class GaReadinessIntegrationTest {
         File(projectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {
@@ -310,7 +310,7 @@ class GaReadinessIntegrationTest {
         File(projectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {
@@ -351,7 +351,7 @@ class GaReadinessIntegrationTest {
         File(projectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kmpFlavors {