+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/GenerateVariantRunConfigurationsTaskTest.kt
@@ -65,7 +65,7 @@ class GenerateVariantRunConfigurationsTaskTest {
         buildFile.writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin {