+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/IntermediateBuildTypeSourceSetTest.kt
@@ -54,7 +54,7 @@ class IntermediateBuildTypeSourceSetTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin {
@@ -125,7 +125,7 @@ class IntermediateBuildTypeSourceSetTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin {