+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/MatrixVariantDependencyInheritanceTest.kt
@@ -52,7 +52,7 @@ class MatrixVariantDependencyInheritanceTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin {