+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/PerVariantDependencyClasspathTest.kt
@@ -58,7 +58,7 @@ class PerVariantDependencyClasspathTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }