+++ build-logic/flavor-plugin/src/test/kotlin/com/mobilebytelabs/kmpflavors/Phase0AutoDetectionTest.kt
@@ -154,7 +154,7 @@ class Phase0AutoDetectionTest {
         File(testProjectDir, "build.gradle.kts").writeText(
             """
             plugins {
-                kotlin("multiplatform") version "2.2.21"
+                kotlin("multiplatform") version "${TestToolchainVersions.kgp}"
                 id("io.github.mobilebytelabs.kmp-product-flavors")
             }
             kotlin {